extracting recurring transactions list fragment into a separate file
