+++ AndroidManifest.xml
@@ -235,9 +235,9 @@
             <action android:name="android.intent.action.EDIT"/>
             <action android:name="android.intent.action.INSERT"/>
         </activity>
-        <activity android:name=".RepeatingTransactionListActivity"/>
+        <activity android:name=".recurring.transactions.RepeatingTransactionListActivity"/>
         <activity
-            android:name=".RepeatingTransactionActivity"
+            android:name=".recurring.transactions.RepeatingTransactionActivity"
             android:label="@string/new_edit_repeating_transaction">
             <action android:name="android.intent.action.EDIT"/>
             <action android:name="android.intent.action.INSERT"/>
